Cleanup: move navbar into common/Navbar
Additionally common/LayoutComponents/LoggedIn has been moved into
Navbar as that is the only use site. Additionally this makes sure that
the LayoutComponents module is fully clean, independent and does not
depend on any of the App's logic code.
