+++ ui/src/components/common/LayoutComponents/index.ts
@@ -1,6 +1,5 @@
 export * from './Column';
 export * from './Container';
-export * from './LoggedIn';
 export * from './HorizontalSeparator';
 export * from './Row';
 export * from './Visible';
