+++ ui/src/components/common/Navbar/NavLinks.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { NavLink } from 'react-router';
 import { twMerge } from 'tailwind-merge';
-import { isNavigationRoute, routeDetails } from '../../router/routeDetails';
+import { isNavigationRoute, routeDetails } from '../../../router/routeDetails';
 
 const testIds = {
   navLink: (translationKey?: string) => `NavLinks::${translationKey}`,
