+++ ui/src/components/common/Navbar/NavigationDropdownMenu.tsx
@@ -1,6 +1,6 @@
 import { Menu, MenuButton, MenuItems } from '@headlessui/react';
 import { FC, ReactNode } from 'react';
-import { dropdownMenuStyles } from '../common/Dropdowns';
+import { dropdownMenuStyles } from '../Dropdowns';
 
 type NavigationDropdownMenuProps = {
   readonly buttonContent: ReactNode;
