+++ ui/src/components/timetables/import/SummarySection.tsx
@@ -1,7 +1,7 @@
 import { FC, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { MdWarning } from 'react-icons/md';
-import { Visible } from '../../common/LayoutComponents/Visible';
+import { Visible } from '../../common/LayoutComponents';
 import { DuplicateJourneysSection } from './duplicate-journeys';
 import { VehicleJourneyDuplicate, VehicleScheduleFrameInfo } from './hooks';
 import { MissingRouteDeviationsSection } from './missing-route-deviations';
