+++ ui/src/router/Router.tsx
@@ -19,11 +19,11 @@ import { getUserInfo } from '../api/user';
 import { PageTitle } from '../components/common';
 import { TaskListDisplay } from '../components/common/AsyncTaskList';
 import { JoreErrorModal } from '../components/common/Modals';
+import { Navbar } from '../components/common/Navbar';
 import { NavigationBlocker } from '../components/forms/common/NavigationBlocker';
 import { MainPage } from '../components/main/MainPage';
 import { MapPage } from '../components/map';
 import { JoreLoader } from '../components/map/JoreLoader';
-import { Navbar } from '../components/navbar';
 import { CreateNewLinePage } from '../components/routes-and-lines/create-line/CreateNewLinePage';
 import { EditLinePage } from '../components/routes-and-lines/edit-line/EditLinePage';
 import { EditRoutePage } from '../components/routes-and-lines/edit-route/EditRoutePage';
